fix: code review 반영
