+++ src/util/decorators/dto-swagger.decorator.ts
@@ -1,0 +1,22 @@
+import {
+    applyDecorators,
+} from "@nestjs/common";
+import {
+    ApiBody, ApiConsumes,
+} from "@nestjs/swagger";
+
+export const FileUploadSwaggerDecorator = () => {
+    return applyDecorators(
+        ApiConsumes("multipart/form-data"), ApiBody({
+            schema: {
+                type: "object",
+                properties: {
+                    file: {
+                        type: "string",
+                        format: "binary",
+                    },
+                },
+            },
+        })
+    );
+};
